changed Direction methods to constants instead
Reasoning behind the change: this allows the type checker to make smarter inferences about directions. For example, if you call `Directions.opposite` method on a value of type `"right" | "left"`, the old version would result in a value of type `Direction`, while the new version will result in the narrower type `"right" | "left"`. This will become especially important once I add more support for diagonal directions, since in the previous system this would affect the return types of all methods and would therefore break code that relied on a method returning a `Direction` as opposed to a `Direction | Diagonal`.

Note: the type-checking methods (e.g. `isDirection`, `isHorizonal`, etc.) have been left unchanged because for these, a method results in better type-checking than a constant.

Note on backwards-compatibility: this breaks existing code, but it does so very loudly and results in errors at compile-time, so it should be easy to fix.